+++ graphistry/__init__.py
@@ -42,6 +42,7 @@ from graphistry.pygraphistry import (  # noqa: E402, F401
     ArrowUploader,
     ArrowFileUploader,
     PyGraphistry,
+    from_igraph
 )
 
 from graphistry.compute import (