+++ graphistry/plugins/__init__.py
@@ -1,0 +1,1 @@
+from .igraph import from_igraph, to_igraph